pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "i-hate-time-tracking"
version = "0.1.0"
description = "Get time tracking out of your way."
authors = ["Dominik Kozaczko <dominik@kozaczko.info>"]
license = "Apache License 2.0"
readme = "README.md"
repository = "https://github.com/dekoza/i-hate-time-tracking"
homepage = "https://github.com/dekoza/i-hate-time-tracking"
keywords = []
packages = [ { include = "ihtt", from = "src" } ]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.4.0"
fastapi = "^0.68.1"
rich = "^10.11.0"
platformdirs = "^2.4.0"
pendulum = "^2.1.2"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
trio = "^0.19.0"
orjson = "^3.6.4"
tortoise-orm = {version = "^0.17.7", extras = ["aiosqlite"]}
aerich = "^0.5.8"
requests = "^2.26.0"

[tool.poetry.group.dev.dependencies]
# formatting, quality, tests
autoflake = "^1.4"
black = "^21.9-beta.0"
isort = "^5.9.3"
mypy = "^0.910"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
pytest-randomly = "^3.10.1"
pytest-sugar = "^0.9.4"
pytest-xdist = "^2.2.0"

# tasks
duty = "^0.7.0"
git-changelog = "^0.4.2"
httpx = "^0.19"
jinja2-cli = "^0.7.0"
toml = "^0.10.2"

# flake8 plugins
darglint = "^1.8.0"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.9.2"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.6.1"
flake8-docstrings = "^1.6.0"
flake8-pytest-style = "^1.5.0"
flake8-string-format = "^0.3.0"
flake8-tidy-imports = "^4.4.1"
flake8-variables-names = "^0.0.4"
pep8-naming = "^0.12.1"
wps-light = "^0.15.3"

# docs
mkdocs = "^1.2.2"
mkdocs-coverage = "^0.2.4"
mkdocs-macros-plugin = "^0.5.12"
mkdocs-material = "^6.2.8"
mkdocstrings = "^0.15.2"

#other
pre-commit = "^2.15.0"
ipython = "^7.28.0"
ipdb = "^0.13.9"


[tool.poetry.scripts]
ih = "ihtt.cli:main"

[tool.black]
exclude = "tests/fixtures"

[tool.isort]
profile = "black"
